refactor: create own package for randomized tests
